chore(file-manager): move secret-manager k8s middleware to common-server
