+++ common-server/pkg/middleware/k8s/suite_test.go
@@ -2,7 +2,7 @@
 //
 // SPDX-License-Identifier: Apache-2.0
 
-package middleware_test
+package k8s_test
 
 import (
 	"testing"
