+++ secret-manager/cmd/server/config/config.go
@@ -5,10 +5,10 @@
 package config
 
 import (
+	"github.com/telekom/controlplane/common-server/pkg/middleware/k8s"
 	"io"
 	"os"
 
-	"github.com/telekom/controlplane/secret-manager/pkg/middleware"
 	"gopkg.in/yaml.v3"
 )
 
@@ -35,10 +35,10 @@ func (c BackendConfig) GetDefault(key string, defaultValue string) string {
 }
 
 type SecurityConfig struct {
-	Enabled        bool                             `yaml:"enabled"`
-	TrustedIssuers []string                         `yaml:"trusted_issuers"`
-	JWKSetURLs     []string                         `yaml:"jwk_set_urls"`
-	AccessConfig   []middleware.ServiceAccessConfig `yaml:"access_config"`
+	Enabled        bool                      `yaml:"enabled"`
+	TrustedIssuers []string                  `yaml:"trusted_issuers"`
+	JWKSetURLs     []string                  `yaml:"jwk_set_urls"`
+	AccessConfig   []k8s.ServiceAccessConfig `yaml:"access_config"`
 }
 
 type ServerConfig struct {
